requirements.txt
Flask==0.10.1
Flask-Assets==0.9
Flask-Login==0.2.11
Flask-Mail==0.9.0
Flask-Principal==0.4.0
Flask-SQLAlchemy==1.0
Flask-Script==2.0.5
Flask-Security==1.7.2
Flask-WTF==0.9.5
Flask-fillin==0.2
Jinja2==2.7.3
Mako==1.0.0
MarkupSafe==0.23
MySQL-python==1.2.5
Pygments==1.6
SQLAlchemy==0.9.4
Sphinx==1.2.2
WTForms==2.0
Werkzeug==0.9.6
alembic==0.6.5
amqp==1.4.5
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.17
blinker==1.3
celery==3.1.11
cssmin==0.2.0
distribute==0.7.3
docutils==0.11
factory-boy==2.3.1
itsdangerous==0.24
jsmin==2.0.9
kombu==3.0.18
lxml==3.3.5
mock==1.0.1
nose==1.3.3
passlib==1.6.2
python-dateutil==2.2
pytz==2014.4
redis==2.10.1
simplejson==3.5.2
six==1.7.2
sphinxcontrib-httpdomain==1.2.1
webassets==0.9
wsgiref==0.1.2
